:art: rename `generic` to `routes`
