+++ app/routes/trust_registry.py
